Temp version for test about CORS

+++ src/main/java/com/sparta/cloneproject_be/service/RoomService.java
@@ -1,28 +1,29 @@
 package com.sparta.cloneproject_be.service;
 
 
+import com.sparta.cloneproject_be.dto.RoomDetailResponseDto;
 import com.sparta.cloneproject_be.dto.RoomRequestDto;
 import com.sparta.cloneproject_be.dto.RoomResponseDto;
+import com.sparta.cloneproject_be.entity.Reservation;
 import com.sparta.cloneproject_be.entity.Room;
 import com.sparta.cloneproject_be.entity.RoomImage;
+import com.sparta.cloneproject_be.entity.User;
 import com.sparta.cloneproject_be.exception.CustomException;
 import com.sparta.cloneproject_be.exception.ErrorMessage;
 import com.sparta.cloneproject_be.repository.ImageRepository;
+import com.sparta.cloneproject_be.repository.ReserveRepository;
 import com.sparta.cloneproject_be.repository.RoomRepository;
-import com.sparta.cloneproject_be.util.S3Uploader;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.bind.annotation.PathVariable;
 
-import java.io.IOException;
+import java.time.Duration;
+import java.time.LocalDate;
 import java.util.*;
-// import static com.sparta.cloneproject_be.exception.ErrorMessage.*;
 
 @Service
 @Slf4j
@@ -31,21 +32,24 @@ public class RoomService {
 
     private final RoomRepository roomRepository;
     private final ImageRepository imageRepository;
+    private final ReserveRepository reserveRepository;
 
     //숙소 게시글 등록
     @Transactional
-    public ResponseEntity<RoomResponseDto> createPost(RoomRequestDto requestDTO, List<String> imgPaths) {
+    public ResponseEntity<RoomResponseDto> createPost(RoomRequestDto requestDTO, List<String> imgPaths, User user) {
         if(!imgPaths.isEmpty()) {
             Room room = new Room(requestDTO);
             roomRepository.save(room);
 
-            List<String> imgList = new ArrayList<>();
+            List<RoomImage> imgList = new ArrayList<>();
             for (String imgUrl : imgPaths) {
                 RoomImage roomImage = new RoomImage(imgUrl, room);
                 imageRepository.save(roomImage);
-                imgList.add(roomImage.getImageUrl());
+                imgList.add(roomImage);
             }
-            return ResponseEntity.status(HttpStatus.CREATED).body(new RoomResponseDto(room));
+            room.setImages(imgList);
+            RoomResponseDto responseDto = new RoomResponseDto(room);
+            return ResponseEntity.status(HttpStatus.CREATED).body(responseDto);
         } else {
             throw new CustomException(ErrorMessage.WRONG_INPUT_IMAGE);
         }
@@ -67,13 +71,14 @@ public class RoomService {
 
     //숙소 게시글 수정
     @Transactional
-    public ResponseEntity<RoomResponseDto> updatePost(Long roomId, RoomRequestDto requestDTO) {
-        // 게시글 존재 유무체크
+    public ResponseEntity<RoomResponseDto> updatePost(Long roomId, RoomRequestDto requestDTO, User user) {
+
+        // 게시글이 존재하는지 확인
         Room room = isRoomExist(roomId);
-        // 게시글 작성자와, 유저 매치 체크
-//        if(checkAuthorIdMatch(room)){
-//            throw new CustomException(WRITER_ONLY_MODIFY);
-//        }
+        // 게시글 작성자와 수정하려는 사용자가 같은치 체크
+        if(checkAuthorIdMatch(room, user)){
+            throw new CustomException(ErrorMessage.CANNOT_UPDATE_POST);
+        }
 
         room.update(requestDTO);
         return ResponseEntity.status(HttpStatus.OK).body(new RoomResponseDto(room));
@@ -81,30 +86,68 @@ public class RoomService {
 
     //숙소 게시글 삭제
     @Transactional
-    public ResponseEntity<String> deletePost(@PathVariable Long roomId) {
-        // 게시글 존재 유무체크
+    public ResponseEntity<String> deletePost(@PathVariable Long roomId, User user) {
+        // 게시글이 존재하는지 확인
         Room room = isRoomExist(roomId);
-        // 게시글 작성자와, 유저 매치 체크
-//        if(checkAuthorIdMatch(post, user)) {
-//            throw new CustomException(WRITER_ONLY_DELETE);
-//        }
+        // 게시글 작성자와 삭제하려는 사용자가 같은지 체크
+        if(checkAuthorIdMatch(room, user)) {
+            throw new CustomException(ErrorMessage.CANNOT_DELETE_POST);
+        }
         roomRepository.delete(room);
         return ResponseEntity.status(HttpStatus.OK).body("게시글 식제 성공");
     }
 
+    // 숙소 상세 페이지 데이터
+    public ResponseEntity<RoomDetailResponseDto> getRoomDetails(Long roomId) {
+        Room room = isRoomExist(roomId);
+        List<LocalDate> enableDates = getEnableDates(room).stream().toList();
+        RoomDetailResponseDto roomDetailResponseDto = new RoomDetailResponseDto(room, enableDates);
+
+        return ResponseEntity.ok(roomDetailResponseDto);
+    }
 
     // id를 매개변수로 받아서 id에 대응되는 게시글이 존재하는지 체크하는 메서드
+
     private Room isRoomExist(Long id){
         return roomRepository.findById(id).orElseThrow(
-                () -> new NullPointerException("error")
+                () -> new CustomException(ErrorMessage.NON_EXIST_POST)
         );
     }
 
+    // enableDates 계산기
+    private Set<LocalDate> getEnableDates(Room room) {
+        List<Reservation> reservations = reserveRepository.findAllByRoomAfterDate(room, LocalDate.now());
+        LocalDate expiredDate = room.getExpiredDate();
+
+        Set<LocalDate> enableDates = getBetweenDates(LocalDate.now(), expiredDate); // List 와 Set 의 contains 메소드 차이 확인하기
+
+        // 2중포문 말고 다른 방법으로 할 수 있을지 고민해보기
+        for (Reservation reservation : reservations) {
+            Set<LocalDate> datesBetweenInAndOut = getBetweenDates(reservation.getCheckIn(), reservation.getCheckOut());
+            for (LocalDate date : datesBetweenInAndOut) {
+                enableDates.remove(date);
+            }
+        }
+
+        return enableDates;
+    }
+
+    // 두 날짜 사이(경계 포함)의 날짜들을 반환하는 메서드
+    private Set<LocalDate> getBetweenDates(LocalDate from, LocalDate to) {
+        Set<LocalDate> betweenDates = new LinkedHashSet<>();
+
+        int betweenFromTo = (int)Duration.between(from.atStartOfDay(), to.atStartOfDay()).toDays();
+        for (int i = 0; i <= betweenFromTo; i++) {
+            betweenDates.add(from.plusDays(i));
+        }
+        return betweenDates;
+    }
+
     // 게시글과 게시글을 변경하려는 요청을 보낸 유저가 일치하는지 여부 체크 default return value : true
-//    private boolean checkAuthorIdMatch(Room room, User user){
-//        if(room.getUser().getUsername().equals(user.getUsername()))
-//            return false;
-//        return true;
-//    }
+    private boolean checkAuthorIdMatch(Room room, User user){
+        if(room.getUser().getUsername().equals(user.getUsername()))
+            return false;
+        return true;
+    }
 
 }
